refactor: the v1/types.go
split types in the v1/types.go to separate files to maintain consistency with the coding style of the rest of the code
